Python-Project_Barkod_And_QR_Reader
Our aim in this project is to create a QR code and barcode reader. Nowadays, QR codes and barcodes are used for almost everything. For example, when you buy a product in a supermarket, the supermarket's sales system detects which product you purchased with its barcode. Another example, when you want to connect to whatsapp web you can easily do this with QR code.
My teammates in this project: halak0013 (Muhammet Halak) and Ömer Faruk Karakay